pattern reorganization and bug fixes
