+++ source/css/00-variables/colors.css
@@ -4,7 +4,7 @@
           
     --color-red-30: #f33;
 
-    --color-gray-10: #3333;
+    --color-gray-10: #333;
     --color-gray-20: #666;
     --color-gray-30: #999;
     --color-gray-40: #ccc;
