+++ source/css/basics/inputs.css
@@ -19,21 +19,19 @@
     }
 }
 
-.input--text {
-    max-width: 20em;
-}
-
 .input--textarea {
     min-height: 8em;  
     max-width: 30em; 
 }
 
-.input--small {
-    max-width: 10em;
+.input--text,
+.input--select,
+.input--medium {
+    max-width: 20em;
 }
 
-.input--medium {
-    max-width: 15em;
+.input--small {
+    max-width: 10em;
 }
 
 .input--large {
