+++ source/css/basics/labels.css
@@ -19,5 +19,5 @@
 .label--checkbox,
 .label--radio {
     display: inline;
-    margin: .25em 0 0 .5em;
+    margin: 0 0 0 .25em;
 }
