+++ source/patterns/basics/messages/formErrorMessage.js
@@ -4,18 +4,17 @@ import classnames from 'classnames';
 function FormErrorMessage({
     children='An unexpected error occurred. Please try again.', 
     id,
-    cssClass,
-    dataAttrs={}
+    cssClass
 }) {
     const cssClasses = classnames('form-message', 'error-message', cssClass);
 
     return (
         <div 
             id={id} 
-            className={cssClasses} 
-            {...dataAttrs}>
+            className={cssClasses}>
             {children}
         </div>
+    );
 }
 
 export default FormErrorMessage;
