+++ source/patterns/basics/video/video.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Video from './video.js';
 
 export default {
-    title: 'Basics/Videos'
+    title: 'Basics/Graphic Elements/Videos'
 };
 
 export const video = () => (
