+++ source/patterns/components/formFields/radioButtonField.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Label from '../../basics/labels/label';
 import RadioButton from '../../basics/inputs/radioButton';
-import FieldErrorMessage from '../../basics/formMessages/fieldErrorMessage';
+import FieldErrorMessage from '../../basics/messages/fieldErrorMessage';
 
 function RadioButtonField({
     id='radio',
