+++ source/patterns/components/rteContent/rteContent.stories.js
@@ -21,8 +21,8 @@ export const withLeftAlignedImage = () => (
 
 export const withRightAlignedImage = () => (
     <RteContent>
-        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
         <img className="align-right" src={contentImage} alt="Alt text goes here" />
+        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
         <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
         <p>Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
     </RteContent>
@@ -31,7 +31,7 @@ export const withRightAlignedImage = () => (
 export const withCenterAlignedImage = () => (
     <RteContent>
         <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
-        <img className="align-left" src={contentImage} alt="Alt text goes here" />
+        <img className="align-center" src={contentImage} alt="Alt text goes here" />
         <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
         <p>Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
     </RteContent>
