+++ source/patterns/pages/pages.stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Homepage from './homepage.js';
+
+export default {
+    title: 'Pages'
+};
+
+export const homepage = () => (
+    <Homepage />
+);
